+++ packages/gestalt/src/Container.js
@@ -3,14 +3,16 @@ import type { Node } from 'react';
 import Box from './Box.js';
 
 type Props = {|
+  /**
+   *
+   */
   children?: Node,
 |};
 
 /**
  * [Containers](https://gestalt.pinterest.systems/container ) are useful in responsively laying out content on different screens.
  */
-export default function Container(props: Props): Node {
-  const { children } = props;
+export default function Container({ children }: Props): Node {
   return (
     <Box justifyContent="center" display="flex">
       <Box maxWidth={800} width="100%">